requirements.txt
pandas
python-dotenv
azure-storage-blob
nbformat
nbconvert
apache-airflow